+++ lightly_studio/tests/resolvers/video/video_frame_resolver/test_get_all_by_dataset_id.py
@@ -11,8 +11,7 @@ from lightly_studio.resolvers import (
 from tests.helpers_resolvers import (
     create_dataset,
 )
-from tests.resolvers.video_frame_resolver.helpers import create_video_with_frames
-from tests.resolvers.video_resolver.helpers import VideoStub
+from tests.resolvers.video.helpers import VideoStub, create_video_with_frames
 
 
 def test_get_all_by_dataset_id(test_db: Session) -> None:
@@ -211,3 +210,35 @@ def test_get_all_by_dataset_id__with_sample_ids(
     assert len(result.samples) == len(sample_ids)
     assert result.total_count == len(sample_ids)
     assert all(sample_id in returned_sample_ids for sample_id in sample_ids)
+
+
+def test_get_all_by_dataset_id__with_video_id(test_db: Session) -> None:
+    dataset = create_dataset(session=test_db, sample_type=SampleType.VIDEO)
+    dataset_id = dataset.dataset_id
+    video_frames = create_video_with_frames(
+        session=test_db,
+        dataset_id=dataset_id,
+        video=VideoStub(path="video1.mp4", duration_s=1, fps=2),
+    )
+
+    create_video_with_frames(
+        session=test_db,
+        dataset_id=dataset_id,
+        video=VideoStub(path="video2.mp4", duration_s=1, fps=2),
+    )
+
+    sample_video_id, video_frames_dataset_id = (
+        video_frames.video_sample_id,
+        video_frames.video_frames_dataset_id,
+    )
+
+    result = video_frame_resolver.get_all_by_dataset_id(
+        session=test_db, dataset_id=video_frames_dataset_id, video_id=sample_video_id
+    )
+
+    assert len(result.samples) == 2
+    assert result.total_count == 2
+    assert result.samples[0].frame_number == 0
+    assert result.samples[0].parent_sample_id == sample_video_id
+    assert result.samples[1].frame_number == 1
+    assert result.samples[1].parent_sample_id == sample_video_id