+++ lightly_studio_view/src/lib/hooks/useClassifiers/useClassifierUtils.ts
@@ -138,7 +138,7 @@ export function useClassifierUtils(): UseClassifierUtilsReturn {
         try {
             const response = await client.POST('/api/classifiers/get_negative_samples', {
                 body: {
-                    dataset_id: datasetId.toString(),
+                    dataset_id: datasetId!.toString(),
                     positive_sample_ids: positives
                 }
             });