+++ lightly_studio_view/src/lib/schema.d.ts
@@ -1224,7 +1224,7 @@ export interface paths {
          *         session: The database session.
          *         video_frame_dataset_id: The ID of the dataset to retrieve frames for.
          *         pagination: Pagination parameters including offset and limit.
-         *
+         *         video_id: The video ID of the frames to retrieve
          *     Returns:
          *         A list of frames along with the total count.
          */
@@ -2281,6 +2281,11 @@ export interface components {
              * @default []
              */
             captions: unknown[];
+            /**
+             * Annotations
+             * @default []
+             */
+            annotations: unknown[];
         };
         /**
          * SampleViewsWithCount
@@ -2553,7 +2558,7 @@ export interface components {
             /** Height */
             height: number;
             /** Duration S */
-            duration_s: number;
+            duration_s?: number | null;
             /** Fps */
             fps: number;
             /** File Name */
@@ -2567,11 +2572,7 @@ export interface components {
             sample_id: string;
             /** Sample */
             sample: unknown;
-            /**
-             * Frames
-             * @default []
-             */
-            frames: components["schemas"]["FrameView"][];
+            frame?: components["schemas"]["FrameView"] | null;
         };
         /**
          * VideoViewsWithCount
@@ -4731,6 +4732,7 @@ export interface operations {
     get_all_frames: {
         parameters: {
             query?: {
+                video_id?: string | null;
                 cursor?: number;
                 limit?: number;
             };